+++ Output/Languages/English.xml
@@ -52,7 +52,6 @@
   <string name="TR_CONFIG_LEVEL3">Level 3</string>
   <string name="TR_CONFIG_LEVEL_MAX">Max</string>
   <string name="TR_CONFIG_LIST">List</string>
-  <string name="TR_CONFIG_LIST_RECORD">List and Record</string>
   <string name="TR_CONFIG_DREIDEL">Dreidel</string>
   <string name="TR_CONFIG_TILE_BOARD">Tiles</string>
   <string name="TR_CONFIG_BOOK">Book</string>
