+++ Output/Languages/German.xml
@@ -47,7 +47,6 @@
   <string name="TR_CONFIG_LEVEL3">Level 3</string>
   <string name="TR_CONFIG_LEVEL_MAX">Maximal</string>
   <string name="TR_CONFIG_LIST">Liste</string>
-  <string name="TR_CONFIG_LIST_RECORD">Liste & Record</string>
   <string name="TR_CONFIG_DREIDEL">Dreidel</string>
   <string name="TR_CONFIG_TILE_BOARD">Kacheln</string>
   <string name="TR_CONFIG_BOOK">Book</string>
