+++ VocaluxeLib/Enums.cs
@@ -201,7 +201,6 @@ namespace VocaluxeLib
     {
         // ReSharper disable InconsistentNaming
         TR_CONFIG_LIST,       // list of covers and artist - title
-        TR_CONFIG_LIST_RECORD,
         //TR_CONFIG_DREIDEL,  //as in ultrastar deluxe
         TR_CONFIG_TILE_BOARD, //chessboard like
         //TR_CONFIG_BOOK      //for playlists
