+++ VocaluxeLib/Menu/SongMenu/CSongMenuFactory.cs
@@ -31,12 +31,6 @@ namespace VocaluxeLib.Menu.SongMenu
                 //case ESongMenu.TR_CONFIG_DREIDEL:
                 //    _SongMenu = new CSongMenuDreidel();
                 //    break;
-                case ESongMenu.TR_CONFIG_LIST_RECORD:
-                    return new CSongMenuListRecord(theme, partyModeID);
-
-                //case ESongMenu.TR_CONFIG_DREIDEL:
-                //    _SongMenu = new CSongMenuDreidel();
-                //    break;
                 case ESongMenu.TR_CONFIG_TILE_BOARD:
                     return new CSongMenuTileBoard(theme, partyModeID);
